+++ preferences/keys.js
@@ -283,6 +283,27 @@ export const SettingsKeys = () => {
       widget_type: 'color',
       themed: true,
     },
+    'icon-border-radius': {
+      default_value: 0,
+      widget_type: 'scale',
+      themed: true,
+    },
+    'icon-border-thickness': {
+      default_value: 0,
+      widget_type: 'dropdown',
+      test: { values: [0, 1, 2, 3] },
+      themed: true,
+    },
+    'icon-border-color': {
+      default_value: [1, 1, 1, 1],
+      widget_type: 'color',
+      themed: true,
+    },
+    'icon-background-color': {
+      default_value: [1, 1, 1, 0],
+      widget_type: 'color',
+      themed: true,
+    },
     'dock-padding': {
       default_value: 0.5,
       widget_type: 'scale',
@@ -427,6 +448,14 @@ export const SettingsKeys = () => {
       default_value: 0,
       widget_type: 'dropdown',
       test: { pointer: 'slide-through', values: [0, 1, 2] },
+    },
+    'overview-transparent-background': {
+      default_value: false,
+      widget_type: 'switch',
+    },
+    'overview-transparent-topbar-background': {
+      default_value: false,
+      widget_type: 'switch',
     },
   });